requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==4.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
requests==2.26.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.7
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
kombu==5.2.3
Markdown==3.3.6
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
